// src/models/movieModel.ts
export interface Movie {
    id: number
    title: string
    year: number
    genre: string[]
    rating: number
    director: string
    actors: string[]
    plot: string
    poster: string
    trailer: string
    runtime: number
    awards: string
    country: string
    language: string
    boxOffice: string
    production: string
    website: string
}